pyproject.toml
[project]
name = "mulag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.9.1",
    "deepagents>=0.2.7",
    "duckduckgo-search>=8.1.1",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-google-genai>=3.1.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pydantic>=2.12.4",
    "pypdf>=6.3.0",
    "python-dotenv>=1.2.1",
]
